rename custom_atlas to anatomical_atlas
