+++ src/ephys_alignment_gui/anatomical_atlas.py
